requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Streamlit for Demo
streamlit==1.25.0

# Explainability
shap==0.42.1
lime==0.2.0.1

# Data Processing
pyarrow==12.0.1
openpyxl==3.1.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Metrics and Logging
tensorboard==2.13.0
mlflow==2.5.0

# Network Analysis (for network-level detection)
scapy==2.5.0
networkx==3.1

# Utilities
tqdm==4.65.0
joblib==1.3.1
python-dateutil==2.8.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter for Notebooks
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.0
